Move binary_sensor data back into the platform-file
